Enemy random spawning fixed, game over screen, level manager, code optimization and comments

+++ app/src/main/java/com/example/exodus/Utils.java
@@ -3,7 +3,7 @@ package com.example.exodus;
 public class Utils {
 
     //getDistanceBetweenPoints returns the distance between 2d points p1 and p2
-    public static double getDistanceBetweenPoints(double p1x, double p1y, double p2x, double p2y) {
-        return Math.sqrt(Math.pow(p1x - p2x, 2) + Math.pow(p1y - p2y, 2));
+    public static float getDistanceBetweenPoints(float p1x, float p1y, float p2x, float p2y) {
+        return (float)Math.sqrt(Math.pow(p1x - p2x, 2) + Math.pow(p1y - p2y, 2));
     }
 }
